make c abi types wasm-friendly

+++ src/lib/root.zig
@@ -4,11 +4,12 @@ const testing = std.testing;
 pub const languages = @import("languages.zig");
 pub const time = @import("time/time.zig");
 
-pub const CMonth = u8;
-pub const CDay = u16;
+// these are the only types supported officially by wasm
+pub const CMonth = i32;
+pub const CDay = i32;
 pub const CYear = i64;
-pub const CEpoch = u64;
-pub const CLanguage = u8;
+pub const CEpoch = i64;
+pub const CLanguage = i32;
 
 comptime {
     _ = time.hekenic;
